Patch YARD::Handlers::Ruby::ClassHandler to recognize Sequel::Model()
Using Sequel::Model(:table)-style superclas definition (the recommended
way of specify a table for a model) normally causes an undocumentable
superclass. This change patches the ClassHandler#parse_superclass method
to handle this in a limited scenario.

Specifically, When it is called exactly as Sequel::Model(...), it will
be recognized. If it is in any other namespace, for example, it will
not.